[CHERRY-PICK] Adding '2268a6413f5de5a299c9ac9db887c849a82f8bb9' to 'CHERRY_PICK_main_2024-02-22_16.26.31' using the python script version 'v1.0.6'.

+++ PresentationLayer/ACCOUNT/AccountingVouchers.aspx.cs
@@ -11935,7 +11935,7 @@ public partial class AccountingVouchers : System.Web.UI.Page
                             objCostCenter.CATID = Convert.ToInt32(CatID);
                             objCostCenter.VCHTYPE = ddlTranType.SelectedValue.ToString();
                             double CCAmount = Convert.ToDouble(lblAmount.Text);
-                            objCostCenter.AMOUNT = Convert.ToInt32(CCAmount);
+                            objCostCenter.AMOUNT = Convert.ToDouble(CCAmount);
 
                             if (isTempVoucher == "Y")
                             {
@@ -17233,7 +17233,7 @@ public partial class AccountingVouchers : System.Web.UI.Page
         dtotheractivity.Columns.Add(new DataColumn("CostCenterID", typeof(int)));
         dtotheractivity.Columns.Add(new DataColumn("PartyID", typeof(int)));
         dtotheractivity.Columns.Add(new DataColumn("PartyName", typeof(string)));
-        dtotheractivity.Columns.Add(new DataColumn("Amount", typeof(decimal)));
+        dtotheractivity.Columns.Add(new DataColumn("Amount", typeof(double)));
 
         return dtotheractivity;
     }
